Typing Test in Python
Copy the text you want to do typing test with onto the Input.txt file. 
Fire the diwali.py program
It will show the correct answers in yellow and the false ones in red.
Pressing backspace would tell you to start the word from the start of the word with an error.